Body Mass Index (BMI) derived variable

This function creates a harmonized BMI variable. The BMI variable provided by the CCHS calculates BMI using methods that vary across cycles, leading to measurement error when using multiple CCHS cycles. In certain CCHS cycles (2001-2003, 2007+), there are age restrictions in which respondents under the age of 20 and over the age of 64 were not included. Across all CCHS cycles, female respondents who identifed as being pregnant were excluded; and in certain CCHS cycles (2003-2007, 2013-2014), females who did not answer the pregnancy question were coded as NS (not stated) for HWTGBMI. As well, in certain CCHS cycles (2001-2003, 2009-2014), respondents outside certain height and weight ranges (0.914-2.108m for height, 0-260kg for weight) were excluded from HWTGBMI.

BMI_fun() creates a derived variable (HWTGBMI_der) that is harmonized across all CCHS cycles. This function divides weight by the square of height.

BMI_fun(HWTGHTM, HWTGWTK)

Arguments

HWTGHTM
CCHS variable for height (in meters)

HWTGWTK
CCHS variable for weight (in kilograms)

Value

numeric value for BMI in the HWTGBMI_der variable

Details

For HWTGBMI_der, there are no restrictions to age, height, weight, or pregnancy status. While pregnancy was consistent across all CCHS cycles, its variable (MAM_037) was not available in the PUMF CCHS datasets so it could not be harmonized and included into the function.

For any single CCHS survey year, it is appropriate to use the CCHS BMI variable (HWTGBMI) that is also available on cchsflow. HWTGBMI_der is recommended when using multiple survey cycles.

HWTGBMI_der uses the CCHS variables for height and weight that have been transformed by cchsflow. In order to generate a value for BMI across CCHS cycles, height and weight must be transformed and harmonized.

Note

In earlier CCHS cycles (2001 and 2003), height was collected in inches; while in later CCHS cycles (2005+) it was collected in meters. To harmonize values across cycles, height was converted to meters (to 3 decimal points). Weight was collected in kilograms across all CCHS cycles, so no transformations were required in the harmonization process.

Examples

# Using BMI_fun() to create BMI values across cycles
# BMI_fun() is specified in variableDetails.csv along with the CCHS variables and cycles 
# included. 

# To transform the derived BMI variable, use RecWTable() for each CCHS cycle and specify 
# HWTGBMI_der, along with height (HWTGHTM) and weight (HWTGWTK). Then by using bind_rows(), you 
# can combined HWTGBMI_der across cycles.

suppressMessages(library(bllflow))

#> Warning: package ‘tableone’ was built under R version 3.5.2

#> Warning: package ‘DDIwR’ was built under R version 3.5.2

#> Warning: package ‘xml2’ was built under R version 3.5.2

#> Warning: package ‘stringr’ was built under R version 3.5.2

#> Warning: package ‘recipes’ was built under R version 3.5.2

#> Warning: package ‘dplyr’ was built under R version 3.5.2

#> Warning: package ‘sjlabelled’ was built under R version 3.5.2

#> Warning: package ‘haven’ was built under R version 3.5.2

library(cchsflow)
bmi2010 <- RecWTable(dataSource = cchs2010, variableDetails = variableDetails,
datasetName = "cchs2010", variables = c("HWTGHTM", "HWTGWTK", "HWTGBMI_der"))

#> [1] "NOTE: Height is a reported in meters from 2005 CCHS onwards"

head(bmi2010)

#>   HWTGHTM HWTGWTK HWTGBMI_der
#> 1   1.600   56.25    21.97266
#> 2   1.600   54.00    21.09375
#> 3   1.829   83.25    24.88610
#> 4   1.727   99.00    33.19331
#> 5   1.753   99.00    32.21598
#> 6   1.727   69.75    23.38619

bmi2012 <- RecWTable(dataSource = cchs2012, variableDetails = variableDetails, datasetName =
"cchs2012", variables = c("HWTGHTM", "HWTGWTK", "HWTGBMI_der"))

#> [1] "NOTE: Height is a reported in meters from 2005 CCHS onwards"

tail(bmi2012)

#>     HWTGHTM HWTGWTK HWTGBMI_der
#> 195   1.829  108.00    32.28467
#> 196   1.524   70.20    30.22506
#> 197   1.803   84.00    25.83972
#> 198   1.575   63.00    25.39683
#> 199   1.829   85.50    25.55870
#> 200   1.702   74.25    25.63170

combined_bmi <- bind_rows(bmi2010, bmi2012)
head(combined_bmi)

#>   HWTGHTM HWTGWTK HWTGBMI_der
#> 1   1.600   56.25    21.97266
#> 2   1.600   54.00    21.09375
#> 3   1.829   83.25    24.88610
#> 4   1.727   99.00    33.19331
#> 5   1.753   99.00    32.21598
#> 6   1.727   69.75    23.38619

tail(combined_bmi)

#>     HWTGHTM HWTGWTK HWTGBMI_der
#> 395   1.829  108.00    32.28467
#> 396   1.524   70.20    30.22506
#> 397   1.803   84.00    25.83972
#> 398   1.575   63.00    25.39683
#> 399   1.829   85.50    25.55870
#> 400   1.702   74.25    25.63170

# Using BMI_fun() to generate a BMI value with user inputted height and weight values
# BMI_fun() can also generate a value for BMI if you input a value for height and weight. Let's 
# say your height is 170cm (1.7m) and your weight is 50kg, your BMI can be calculated as follows:

library(cchsflow)
BMI <- BMI_fun(HWTGHTM = 1.7, HWTGWTK = 50)
print(BMI)

#> [1] 17.30104