Smoking pack-years

This function creates a derived variable (Pack_years_der) that measures an individual smoking pack-years based on various CCHS smoking variables. This is a popular variable used by researchers to quantify lifetime exposure to cigarette use.

Pack_years_fun(
  SMKDSTY,
  DHHGAGE_cont,
  SMK_09A_B,
  SMKG09C,
  SMKG203_cont,
  SMKG207_cont,
  SMK_204,
  SMK_05B,
  SMK_208,
  SMK_05C,
  SMKG01C_cont,
  SMK_01A
)

Arguments

SMKDSTY
derived variable that classifies an individual's smoking status.

DHHGAGE_cont
continuous age variable.

SMK_09A_B
number of years since quitting smoking. Variable asked to former daily smokers who quit <3 years ago.

SMKG09C
number of years since quitting smoking. Variable asked to former daily smokers who quit >=3 years ago.

SMKG203_cont
age started smoking daily. Variable asked to daily smokers.

SMKG207_cont
age started smoking daily. Variable asked to former daily smokers.

SMK_204
number of cigarettes smoked per day. Variable asked to daily smokers.

SMK_05B
number of cigarettes smoked per day. Variable asked to occasional smokers

SMK_208
number of cigarettes smoked per day. Variable asked to former daily smokers

SMK_05C
number of days smoked at least one cigarette

SMKG01C_cont
age smoked first cigarette

SMK_01A
smoked 100 cigarettes in lifetime (y/n)

Value

value for smoking pack-years in the Pack_years_der variable

Details

pack-years is calculated by multiplying the number of cigarette packs per day (20 cigarettes per pack) by the number of years. Example 1: a respondent who is a current smoker who smokes 1 package of cigarettes for the last 10 years has smoked 10 pack-years. Pack-years is also calculated for former smokers. Example 2: a respondent who started smoking at age 20 years and smoked half a pack of cigarettes until age 40 years smoked for 10 pack-years.

Examples

# Using Pack_years_fun() to create pack-years values across CCHS cycles
# Pack_years_fun() is specified in variableDetails.csv along with the CCHS variables and cycles included. 

# To transform Pack_years_der across cycles, use RecWTable() for each CCHS cycle and specify Pack_years_der, along with
# each smoking variable. Then by using bind_rows(), you can combine Pack_years_der across cycles

suppressMessages(library(bllflow))
library(cchsflow)
pack_years2010 <- RecWTable(dataSource = cchs2010, variableDetails = variableDetails, datasetName = "cchs2010",
variables = c("SMKDSTY", "DHHGAGE_cont", "SMK_09A_B", "SMKG09C", "SMKG203_cont", "SMKG207_cont", "SMK_204", "SMK_05B",
"SMK_208", "SMK_05C", "SMK_01A", "SMKG01C_cont", "Pack_years_der"))
head(pack_years2010)

#>   DHHGAGE_cont SMKDSTY SMKG01C_cont SMK_01A SMKG203_cont SMKG207_cont SMK_204
#> 1           22       6           NA       2           NA           NA      NA
#> 2           42       6           NA       2           NA           NA      NA
#> 3           37       1         18.5       1         18.5           NA      10
#> 4           37       6           NA       2           NA           NA      NA
#> 5           37       6           NA       2           NA           NA      NA
#> 6           32       4         16.0       1           NA           16      NA
#>   SMK_05B SMK_05C SMK_208 SMK_09A_B SMKG09C Pack_years_der
#> 1      NA      NA      NA     NA(a)   NA(a)           0.00
#> 2      NA      NA      NA     NA(a)   NA(a)           0.00
#> 3      NA      NA      NA     NA(a)   NA(a)           9.25
#> 4      NA      NA      NA     NA(a)   NA(a)           0.00
#> 5      NA      NA      NA     NA(a)   NA(a)           0.00
#> 6      NA      NA      12         1   NA(a)           9.30

pack_years2012 <- RecWTable(dataSource = cchs2012, variableDetails = variableDetails, datasetName = "cchs2012",
variables = c("SMKDSTY", "DHHGAGE_cont", "SMK_09A_B", "SMKG09C", "SMKG203_cont", "SMKG207_cont", "SMK_204", "SMK_05B",
"SMK_208", "SMK_05C", "SMK_01A", "SMKG01C_cont", "Pack_years_der"))
tail(pack_years2012)

#>     DHHGAGE_cont SMKDSTY SMKG01C_cont SMK_01A SMKG203_cont SMKG207_cont SMK_204
#> 195           42       6           NA       2           NA           NA      NA
#> 196           57       6           NA       2           NA           NA      NA
#> 197           32       4          8.0       1           NA           13      NA
#> 198           77       5         18.5       2           NA           NA      NA
#> 199           47       1         16.0       1           16           NA       5
#> 200           85       5          8.0       2           NA           NA      NA
#>     SMK_05B SMK_05C SMK_208 SMK_09A_B SMKG09C Pack_years_der
#> 195      NA      NA      NA     NA(a)   NA(a)          0.000
#> 196      NA      NA      NA     NA(a)   NA(a)          0.000
#> 197      NA      NA      20         3   NA(a)         16.500
#> 198      NA      NA      NA     NA(a)   NA(a)          0.007
#> 199      NA      NA      NA     NA(a)   NA(a)          7.750
#> 200      NA      NA      NA     NA(a)   NA(a)          0.007

combined_pack_years <- bind_rows(pack_years2010, pack_years2012)

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

head(combined_pack_years)

#>   DHHGAGE_cont SMKDSTY SMKG01C_cont SMK_01A SMKG203_cont SMKG207_cont SMK_204
#> 1           22       6           NA       2           NA           NA      NA
#> 2           42       6           NA       2           NA           NA      NA
#> 3           37       1         18.5       1         18.5           NA      10
#> 4           37       6           NA       2           NA           NA      NA
#> 5           37       6           NA       2           NA           NA      NA
#> 6           32       4         16.0       1           NA           16      NA
#>   SMK_05B SMK_05C SMK_208 SMK_09A_B SMKG09C Pack_years_der
#> 1      NA      NA      NA     NA(a)   NA(a)           0.00
#> 2      NA      NA      NA     NA(a)   NA(a)           0.00
#> 3      NA      NA      NA     NA(a)   NA(a)           9.25
#> 4      NA      NA      NA     NA(a)   NA(a)           0.00
#> 5      NA      NA      NA     NA(a)   NA(a)           0.00
#> 6      NA      NA      12         1   NA(a)           9.30

tail(combined_pack_years)

#>     DHHGAGE_cont SMKDSTY SMKG01C_cont SMK_01A SMKG203_cont SMKG207_cont SMK_204
#> 395           42       6           NA       2           NA           NA      NA
#> 396           57       6           NA       2           NA           NA      NA
#> 397           32       4          8.0       1           NA           13      NA
#> 398           77       5         18.5       2           NA           NA      NA
#> 399           47       1         16.0       1           16           NA       5
#> 400           85       5          8.0       2           NA           NA      NA
#>     SMK_05B SMK_05C SMK_208 SMK_09A_B SMKG09C Pack_years_der
#> 395      NA      NA      NA     NA(a)   NA(a)          0.000
#> 396      NA      NA      NA     NA(a)   NA(a)          0.000
#> 397      NA      NA      20         3   NA(a)         16.500
#> 398      NA      NA      NA     NA(a)   NA(a)          0.007
#> 399      NA      NA      NA     NA(a)   NA(a)          7.750
#> 400      NA      NA      NA     NA(a)   NA(a)          0.007