Participation and Activity Limitation

This is a derived variable used in the CCHS (RACDPAL) to classify respondents according to the frequency with which they experience activity limitations due to disability.

RACDPAL_fun(RAC_1, RAC_2A, RAC_2B, RAC_2C)

Arguments

RAC_1

Has difficulty with activities due to disability

RAC_2A

Reduction in activities at home due to disability

RAC_2B

Reduction in activities at school or work due to disability

RAC_2C

Reduction in other activities

Value

the CCHS derived variable RACDPAL with 3 categories:

Sometimes

Often

Never

Details

This derived variable is generated in CCHS cycles 2003-2014. The 2001 CCHS cycle, however, contains the same base variables used to derive this variable. To include respondents in the 2001 CCHS cycle, this custom function was created using the same derivation conditions used in later cycles.

Examples

# Using RACDPAL_fun() to transform RACDPAL in 2001. 
# RACDPAL_fun() is specified in variable_details.csv along with the
# CCHS variables and cycles included.
 
# To transform RACDPAL, use rec_with_table() for each the 2001 cycle
# and specify RACDPAL, along with the various ADL variables.

library(cchsflow)

RACDPAL_2001 <- rec_with_table(
  cchs2001_p, c(
    "RAC_1", "RAC_2A", "RAC_2B", "RAC_2C", "RACDPAL"
  )
)
#> No variable_details detected.
#>               Loading cchsflow variable_details
#> Using the passed data variable name as database_name
#> NOTE for RAC_2B: In CCHS 2001, work and school reduction were combined into one variable. From CCHS 2003 onwards they were split into 2 variables (RAC_2B1 & RAC_2B2)

head(RACDPAL_2001)
#>   RAC_1 RAC_2A RAC_2B RAC_2C RACDPAL
#> 1     3      3  NA(a)      3   NA(b)
#> 2     2      2  NA(a)      2   NA(b)
#> 3     3      3      3      3       3
#> 4     1      3  NA(a)      3   NA(b)
#> 5     3      3      3      3       3
#> 6     3      3  NA(a)      3   NA(b)

# Note: In other CCHS cycles you only need to specify RACDPAL as the variable
# was included in those survey cycles.

# Using RACDPAL_fun() with user inputted data.

# Let's say you're an individual that sometimes has difficulties with
# activities due to disability, sometimes has a reduction in activities at
# home, often has a reduction at school or work, and never has a reduction
# in other activities. Your participation and activity limitation can be
# determined as follows:

library(cchsflow)
RACDPAL <- RACDPAL_fun(1, 1, 2, 3)
print(RACDPAL)
#> [1] 2