Resp_condition_fun3

This is one of 3 functions used to create a derived variable (Resp_condition_der) that determines if a respondents has a respirtory condition. This function for CCHS cycles (2001-2003) that use COPD and Emphysema as a combined variable, as well as Bronchitis

RespCondition_fun3(DHHGAGE_cont, CCC_091, CCC_91A)

Arguments

DHHGAGE_cont
continuous age variable.

CCC_091
variable indicating if respondent has either COPD or Emphysema

CCC_91A
variable indicating if respondent has chronic bronchitis cchsflow variable name: CCC_91A

Value

a categorical variable (Resp_condition_der) with 3 levels:

1 - respondent is over the age of 35 and has a respiratory condition

2 - respondent is under the age of 35 and has a respiratory condition

3 - respondent does not have a respiratory condition