Resp_condition_fun2

This is one of 3 functions used to create a derived variable (Resp_condition_der) that determines if a respondents has a respirtory condition. This function is for CCHS cycles (2005-2007) that use COPD & Emphysema as separate variables, as well as Bronchitis.

Resp_condition_fun2(DHHGAGE_cont, CCC_91E, CCC_91F, CCC_91A)

Arguments

DHHGAGE_cont
continuous age variable.

CCC_91E
variable indicating if respondent has emphysema

CCC_91F
variable indicating if respondent has COPD

CCC_91A
variable indicating if respondent has chronic bronchitis

Value

a categorical variable (Resp_condition_der) with 3 levels:

1 - respondent is over the age of 35 and has a respiratory condition

2 - respondent is under the age of 35 and has a respiratory condition

3 - respondent does not have a respiratory condition

See also

Examples

# Using Resp_condition_fun2() to create pack-years values across CCHS cycles (2005-2007)
# Resp_condition_fun2() is specified in variableDetails.csv along with the CCHS variables and cycles included.

# To transform Resp_condition_der, use RecWTable() for each CCHS cycle and specify Resp_condition_der, 
# along with the various respiratory variables. Then by using bind_rows(), you can combined Resp_condition_der
# across cycles.

suppressMessages(library(bllflow))
library(cchsflow)
resp2005 <- RecWTable(dataSource = cchs2005, variableDetails = variableDetails, datasetName = "cchs2005",
variables = c("DHHGAGE_cont", "CCC_91E", "CCC_91F", "CCC_91A", "Resp_condition_der"))
head(resp2005)

#>   DHHGAGE_cont CCC_91A CCC_91E CCC_91F Resp_condition_der
#> 1           27       2   NA(a)   NA(a)                 NA
#> 2           77       2       2       2                  3
#> 3           32       2       2       2                  3
#> 4           77       2       2       2                  3
#> 5           85       2       2       2                  3
#> 6           13       2   NA(a)   NA(a)                 NA

resp2007_2008 <- RecWTable(dataSource = cchs2007_2008, variableDetails = variableDetails, datasetName = "cchs2007_2008",
variables = c("DHHGAGE_cont", "CCC_91E", "CCC_91F", "CCC_91A", "Resp_condition_der"))
tail(resp2007_2008)

#>     DHHGAGE_cont CCC_91A CCC_91E CCC_91F Resp_condition_der
#> 195           62       2       2       2                  3
#> 196           37       2       2       2                  3
#> 197           77       2       2       2                  3
#> 198           42       2       2       2                  3
#> 199           32       2       2       2                  3
#> 200           72   NA(b)       2       2                 NA

combined_resp <- bind_rows(resp2005, resp2007_2008)

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

head(combined_resp)

#>   DHHGAGE_cont CCC_91A CCC_91E CCC_91F Resp_condition_der
#> 1           27       2   NA(a)   NA(a)                 NA
#> 2           77       2       2       2                  3
#> 3           32       2       2       2                  3
#> 4           77       2       2       2                  3
#> 5           85       2       2       2                  3
#> 6           13       2   NA(a)   NA(a)                 NA

tail(combined_resp)

#>     DHHGAGE_cont CCC_91A CCC_91E CCC_91F Resp_condition_der
#> 395           62       2       2       2                  3
#> 396           37       2       2       2                  3
#> 397           77       2       2       2                  3
#> 398           42       2       2       2                  3
#> 399           32       2       2       2                  3
#> 400           72   NA(b)       2       2                 NA