2009 CCHS synthetic subset data (200 respondents)

This is a subset of 200 observations from the 2009 cycle of the Canadian Community Health Survey (CCHS) synthetic dataset. The CCHS survey is conducted by Statistics Canada.

Source

https://www.statcan.gc.ca/en/statistical-programs/document/3226_D56_T9_V1

Value

cchs2009_s

a data frame

Details

NOTE: this subset of respondents may also be in the 2009 synthetic subset. Please see the "CCHS datasets that overlap each other" article to see how the two datasets contain overlap.

See here for the open license. Source from Statistics Canada, Canadian Community Health Survey PUMF, accessed May 2022. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.

Long name: synthetic-CCHS-E-2009-FullSampleFile_F1

DDI: https://osf.io/q8g7y/

Additional documentation (PDFs): https://osf.io/hkuy3/

Examples

data(cchs2009_s)
str(cchs2009_s)
#> 'data.frame':	200 obs. of  1845 variables:
#>  $ VERDATE : int  20100910 20100910 20100910 20100910 20100910 20100910 20100910 20100910 20100910 20100910 ...
#>  $ REFPER  : chr  "200901-200912" "200901-200912" "200901-200912" "200901-200912" ...
#>  $ SAMPLEID: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ PERSONID: logi  NA NA NA NA NA NA ...
#>  $ GEO_PRV : int  10 10 10 10 10 10 10 10 10 10 ...
#>  $ GEODPC  : chr  "A2N1K7" "A2N2K1" "A0J1S0" "A1V1S7" ...
#>  $ GEODHR4 : int  1013 1013 1012 1012 1012 1011 1011 1012 1011 1014 ...
#>  $ GEODSHR : int  999996 999996 999996 999996 999996 999996 999996 999996 999996 999996 ...
#>  $ GEODDHA : int  996 996 996 996 996 996 996 996 996 996 ...
#>  $ GEODLHA : int  996 996 996 996 996 996 996 996 996 996 ...
#>  $ GEODLHN : int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODDA06: int  10040084 10040089 10080213 10060067 10060109 10010661 10010304 10060103 10010630 10100094 ...
#>  $ GEODFED : int  10005 10005 10003 10002 10002 10007 10007 10002 10007 10004 ...
#>  $ GEODCSD : int  1004019 1004019 1008059 1006009 1006017 1001542 1001519 1006017 1001542 1010032 ...
#>  $ GEODCD  : int  1004 1004 1008 1006 1006 1001 1001 1006 1001 1010 ...
#>  $ GEODSAT : int  6 6 7 6 3 1 1 3 1 6 ...
#>  $ GEODCMA6: int  0 0 0 0 0 1 1 0 1 0 ...
#>  $ GEODPG09: int  9 9 9 9 9 3 3 9 3 8 ...
#>  $ GEODUR  : int  4 4 9 4 1 1 1 1 1 4 ...
#>  $ GEODUR2 : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ GEODPSZ : int  4 4 4 4 4 4 4 4 4 4 ...
#>  $ SAM_CP  : logi  NA NA NA NA NA NA ...
#>  $ SAM_TYP : logi  NA NA NA NA NA NA ...
#>  $ SAMDSHR : logi  NA NA NA NA NA NA ...
#>  $ SAMDLNK : logi  NA NA NA NA NA NA ...
#>  $ ADM_STA : logi  NA NA NA NA NA NA ...
#>  $ ADM_PRX : logi  NA NA NA NA NA NA ...
#>  $ ADM_YOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_MOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_DOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_N09 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N10 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N11 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N12 : logi  NA NA NA NA NA NA ...
#>  $ ADM_LHH : logi  NA NA NA NA NA NA ...
#>  $ DHH_AGE : int  22 24 19 75 62 38 13 24 15 45 ...
#>  $ DHH_YOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_MOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_DOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_SEX : int  1 2 2 1 2 2 2 2 1 1 ...
#>  $ DHH_MS  : int  6 6 6 1 1 1 6 1 6 6 ...
#>  $ DHHDSAGE: int  996 996 996 71 63 51 996 35 996 50 ...
#>  $ DHHDYKD : int  0 0 1 0 0 2 2 0 2 0 ...
#>  $ DHHDOKD : int  0 0 0 0 0 0 0 0 1 0 ...
#>  $ DHHDLE5 : int  0 0 1 0 0 0 0 0 0 0 ...
#>  $ DHHD611 : int  0 0 0 0 0 2 0 0 1 0 ...
#>  $ DHHDL12 : int  0 0 1 0 0 2 0 0 1 0 ...
#>  $ DHHDL18 : int  0 0 1 0 0 2 2 0 3 0 ...
#>  $ DHHDLVG : int  1 8 9 3 3 4 7 3 9 3 ...
#>  $ DHHDECF : int  1 5 5 3 3 5 9 3 5 3 ...
#>  $ DHHDHSZ : int  1 3 4 2 2 4 3 2 7 2 ...
#>  $ GEN_01  : int  2 3 2 3 2 2 2 2 2 2 ...
#>  $ GEN_02  : int  3 1 1 3 3 3 3 3 3 3 ...
#>  $ GEN_02A2: int  5 7 8 7 8 8 8 7 10 8 ...
#>  $ GEN_02B : int  2 4 1 2 3 1 1 2 2 2 ...
#>  $ GEN_07  : int  4 4 2 1 2 3 2 2 2 4 ...
#>  $ GEN_08  : int  1 1 1 2 2 1 6 2 2 1 ...
#>  $ GEN_09  : int  4 3 3 6 6 4 6 6 6 4 ...
#>  $ GEN_10  : int  4 4 2 2 4 2 3 3 2 2 ...
#>  $ GENDHDI : int  3 2 3 2 3 3 3 3 3 3 ...
#>  $ GENDMHI : int  3 1 4 3 2 4 4 3 3 3 ...
#>  $ GENGSWL : int  3 2 2 2 2 2 2 2 1 2 ...
#>  $ DOORG   : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ ORG_1   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ ORG_2   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ DOCIH   : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ CIH_1   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_2   : int  96 96 96 96 96 96 96 96 96 96 ...
#>  $ CIH_3   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_4   : int  96 96 96 96 96 96 96 96 96 96 ...
#>  $ CIH_5   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6A  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6I  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6B  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6J  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6K  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6G  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6F  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6E  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6L  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6M  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6N  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6H  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_7   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8A  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8B  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8C  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8J  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8K  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8G  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8L  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8H  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8I  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ DOHCS   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ HCS_1   : int  9 9 2 4 3 1 6 2 4 7 ...
#>  $ HCS_2   : int  9 9 3 3 1 2 6 3 3 9 ...
#>  $ HCS_3   : int  9 9 2 3 1 2 6 3 4 9 ...
#>  $ HCS_4   : int  9 9 2 3 1 2 6 2 4 9 ...
#>  $ MAM_037 : int  6 2 2 6 6 2 6 2 6 6 ...
#>   [list output truncated]