2010 CCHS synthetic subset data (200 respondents)

This is a subset of 200 observations from the 2010 cycle of the Canadian Community Health Survey (CCHS) synthetic dataset. The CCHS survey is conducted by Statistics Canada.

Source

https://www.statcan.gc.ca/en/statistical-programs/document/3226_D56_T9_V1

Value

cchs2010_s

a data frame

Details

NOTE: this subset of respondents may also be in the 2010 synthetic subset. Please see the "CCHS datasets that overlap each other" article to see how the two datasets contain overlap.

See here for the open license. Source from Statistics Canada, Canadian Community Health Survey PUMF, accessed May 2022. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.

Long name: synthetic-CCHS-E-2010-AnnualComponent_F1

DDI: https://osf.io/q8g7y/

Additional documentation (PDFs): https://osf.io/hkuy3/

Examples

data(cchs2010_s)
str(cchs2010_s)

#> 'data.frame':	200 obs. of  1955 variables:
#>  $ VERDATE : int  20110722 20110722 20110722 20110722 20110722 20110722 20110722 20110722 20110722 20110722 ...
#>  $ REFPER  : Factor w/ 1 level "201001-201012": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ SAMPLEID: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ PERSONID: logi  NA NA NA NA NA NA ...
#>  $ GEO_PRV : int  10 10 10 10 10 10 10 10 10 10 ...
#>  $ GEODPC  : Factor w/ 200 levels "A0G1G0","A0G2S0",..: 167 6 166 79 97 172 164 162 193 42 ...
#>  $ GEODHR4 : int  1013 1012 1013 1011 1012 1014 1013 1013 1014 1011 ...
#>  $ GEODBCHA: int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODSHR : int  999996 999996 999996 999996 999996 999996 999996 999996 999996 999996 ...
#>  $ GEODDHA : int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODRHA : int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODLHA : int  996 996 996 996 996 996 996 996 996 996 ...
#>  $ GEODLHN : int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODDA06: int  10040080 10060125 10040087 10010658 10060063 10100091 10040086 10040090 10100087 10010304 ...
#>  $ GEODFED : int  10005 10002 10005 10007 10002 10004 10005 10005 10004 10007 ...
#>  $ GEODCSD : int  1004019 1006029 1004019 1001542 1006009 1010032 1004019 1004019 1010032 1001519 ...
#>  $ GEODCD  : int  1004 1006 1004 1001 1006 1010 1004 1004 1010 1001 ...
#>  $ GEODSAT : int  6 7 6 1 6 6 6 6 6 1 ...
#>  $ GEODCMA6: int  0 0 0 1 0 0 0 0 0 1 ...
#>  $ GEODPG09: int  9 9 9 3 9 8 9 9 8 3 ...
#>  $ GEODUR  : int  4 9 9 1 0 4 4 4 9 1 ...
#>  $ GEODUR2 : int  1 1 1 1 2 1 1 1 1 1 ...
#>  $ GEODPSZ : int  4 4 4 4 4 4 4 4 4 4 ...
#>  $ SAM_CP  : logi  NA NA NA NA NA NA ...
#>  $ SAM_TYP : logi  NA NA NA NA NA NA ...
#>  $ SAMDSHR : logi  NA NA NA NA NA NA ...
#>  $ SAMDLNK : logi  NA NA NA NA NA NA ...
#>  $ ADM_STA : logi  NA NA NA NA NA NA ...
#>  $ ADM_PRX : logi  NA NA NA NA NA NA ...
#>  $ ADM_YOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_MOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_DOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_N09 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N10 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N11 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N12 : logi  NA NA NA NA NA NA ...
#>  $ ADM_LHH : logi  NA NA NA NA NA NA ...
#>  $ DHH_AGE : int  82 24 63 69 62 48 48 20 58 25 ...
#>  $ DHH_YOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_MOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_DOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_SEX : int  2 1 1 2 1 2 1 1 2 2 ...
#>  $ DHH_MS  : int  1 6 1 5 5 1 6 6 1 1 ...
#>  $ DHHDSAGE: int  91 996 61 996 996 43 996 996 55 37 ...
#>  $ DHHDYKD : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHDOKD : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHDLE5 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHD611 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHDL12 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHDL18 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHDLVG : int  3 1 3 1 1 3 1 2 3 3 ...
#>  $ DHHDECF : int  3 1 3 1 1 3 1 2 3 3 ...
#>  $ DHHDHSZ : int  2 1 2 1 1 2 1 2 2 2 ...
#>  $ GEN_01  : int  4 3 3 3 2 2 2 2 4 2 ...
#>  $ GEN_02  : int  5 3 2 4 3 3 3 1 3 3 ...
#>  $ GEN_02A2: int  9 7 8 5 9 10 8 8 8 9 ...
#>  $ GEN_02B : int  4 1 2 2 2 1 3 1 3 2 ...
#>  $ GEN_07  : int  1 2 3 2 3 2 3 2 2 5 ...
#>  $ GEN_08  : int  6 2 2 2 1 2 1 2 2 2 ...
#>  $ GEN_09  : int  6 6 6 6 3 6 3 6 6 6 ...
#>  $ GEN_10  : int  3 1 3 2 2 1 2 2 1 2 ...
#>  $ GENDHDI : int  1 2 2 2 3 3 3 3 1 3 ...
#>  $ GENDMHI : int  1 4 3 3 3 4 2 4 2 3 ...
#>  $ GENGSWL : int  1 2 2 3 1 1 2 2 2 1 ...
#>  $ DOORG   : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ ORG_1   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ ORG_2   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ DOCIH   : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ CIH_1   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_2   : int  96 96 96 96 96 96 96 96 96 96 ...
#>  $ CIH_3   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_4   : int  96 96 96 96 96 96 96 96 96 96 ...
#>  $ CIH_5   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6A  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6I  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6B  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6J  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6K  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6G  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6F  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6E  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6L  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6M  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6N  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6H  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_7   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8A  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8B  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8C  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8J  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8K  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8G  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8L  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8H  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8I  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ DOHCS   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ HCS_1   : int  3 2 3 4 2 4 2 2 2 2 ...
#>  $ HCS_2   : int  2 2 3 4 2 3 2 1 2 2 ...
#>  $ HCS_3   : int  2 2 4 4 2 3 2 1 2 2 ...
#>   [list output truncated]