2014 CCHS PUMF subset data (200 respondents)

This is a subset of 200 observations from the 2014 cycle of the Canadian Community Health Survey (CCHS) Public Use Microdata file (PUMF) dataset. The CCHS survey is conducted by Statistics Canada.

Source

https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=164081

Value

cchs2014_p

a data frame

Details

NOTE: this subset of respondents may also be in the 2013-2014 PUMF subset. Please see the "CCHS datasets that overlap each other" article to see how the two datasets contain overlap.

See here for the open license. Source from Statistics Canada, Canadian Community Health Survey PUMF, accessed Jan 2020. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.

Long name: cchs-82M0013-E-2014-Annual-component

DDI: https://osf.io/tbmdn/

Additional documentation (PDFs): https://osf.io/hkuy3/

Examples

data(cchs2014_p)
str(cchs2014_p)
#> 'data.frame':	200 obs. of  1129 variables:
#>  $ VERDATE : int  20150326 20150326 20150326 20150326 20150326 20150326 20150326 20150326 20150326 20150326 ...
#>  $ ADM_RNO : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ GEOGPRV : int  46 24 35 35 60 35 60 35 35 35 ...
#>  $ GEODPMF : int  46901 24906 35995 35965 60901 35953 60901 35944 35951 35953 ...
#>  $ GEODBCHA: int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ ADM_PRX : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ ADM_N09 : int  6 1 2 2 9 2 2 6 1 6 ...
#>  $ ADM_N10 : int  1 1 1 1 9 1 1 1 1 1 ...
#>  $ ADM_N11 : int  6 6 6 6 9 6 6 6 6 6 ...
#>  $ DHHGAGE : int  10 12 12 5 11 8 13 14 13 3 ...
#>  $ DHH_SEX : int  2 2 1 2 1 1 1 2 2 1 ...
#>  $ DHHGLE5 : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHG611 : int  0 0 0 0 0 1 0 0 0 0 ...
#>  $ DHHGL12 : int  0 0 0 0 0 1 0 0 0 0 ...
#>  $ DHHGMS  : int  3 4 3 1 4 1 1 1 3 4 ...
#>  $ DHHGLVG : int  1 1 1 3 1 4 3 1 1 7 ...
#>  $ DHHGHSZ : int  1 1 1 2 1 3 2 1 1 4 ...
#>  $ GEN_01  : int  2 3 3 2 4 1 5 3 1 1 ...
#>  $ GEN_02  : int  3 3 3 3 3 2 5 3 3 3 ...
#>  $ GEN_02A2: int  9 7 8 8 5 9 9 7 5 10 ...
#>  $ GEN_02B : int  2 1 2 2 3 1 1 2 4 1 ...
#>  $ GEN_07  : int  2 4 3 3 3 3 5 2 3 1 ...
#>  $ GEN_08  : int  2 1 1 1 2 1 2 2 2 1 ...
#>  $ GEN_09  : int  6 3 3 3 6 2 6 6 6 1 ...
#>  $ GEN_10  : int  4 2 2 2 3 1 2 2 4 2 ...
#>  $ GENDHDI : int  3 2 2 3 1 4 0 2 4 4 ...
#>  $ GENDMHI : int  3 4 3 3 2 4 4 3 1 4 ...
#>  $ GENGSWL : int  1 2 2 2 3 1 1 2 3 1 ...
#>  $ DOSLP   : int  1 2 2 2 2 2 2 2 2 2 ...
#>  $ SLPG01  : int  7 96 96 96 96 96 96 96 96 96 ...
#>  $ SLP_02  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ SLP_03  : int  4 6 6 6 6 6 6 6 6 6 ...
#>  $ SLP_04  : int  1 6 6 6 6 6 6 6 6 6 ...
#>  $ DOCIH   : int  1 2 2 2 2 2 2 2 2 2 ...
#>  $ CIH_1   : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_2   : int  96 96 96 96 96 96 96 96 96 96 ...
#>  $ CIH_3   : int  1 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_4   : int  1 96 96 96 96 96 96 96 96 96 ...
#>  $ CIH_5   : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6A  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6I  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6B  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6J  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6K  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6G  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6F  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6E  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6L  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6M  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6N  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_6H  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_7   : int  1 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8A  : int  1 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8B  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8C  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8J  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8K  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8G  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8L  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8H  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ CIH_8I  : int  2 6 6 6 6 6 6 6 6 6 ...
#>  $ DOOH1   : int  2 2 1 1 2 1 2 1 1 1 ...
#>  $ OH1_20  : int  6 6 3 2 6 1 6 3 2 1 ...
#>  $ OH1_21A : int  6 6 1 1 6 1 6 1 1 1 ...
#>  $ OH1_21B : int  6 6 1 1 6 1 6 1 2 1 ...
#>  $ OH1_21C : int  6 6 1 1 6 1 6 1 1 1 ...
#>  $ OH1_22  : int  6 6 4 4 6 4 6 4 4 4 ...
#>  $ DOHCS   : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ HCS_1   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ HCS_2   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ HCS_3   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ HCS_4   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ HWT_4   : int  3 1 3 6 9 3 1 3 3 3 ...
#>  $ HWTGHTM : num  1.68 1.63 1.68 10 1.65 ...
#>  $ HWTGWTK : num  63 74.2 69 1000 1000 ...
#>  $ HWTGBMI : num  22.4 28.1 24.6 1000 1000 ...
#>  $ HWTGISW : int  2 3 2 9 9 3 4 3 3 2 ...
#>  $ HWTDCOL : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ HWTDWHO : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_031 : int  2 2 2 1 2 2 2 2 2 2 ...
#>  $ CCC_035 : int  6 6 6 1 6 6 6 6 6 6 ...
#>  $ CCC_036 : int  6 6 6 1 6 6 6 6 6 6 ...
#>  $ CCC_041 : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ CCC_051 : int  2 2 1 2 2 2 1 1 1 2 ...
#>  $ CCC_061 : int  2 1 2 2 2 2 2 2 2 2 ...
#>  $ CCC_071 : int  2 1 2 2 2 2 1 1 2 2 ...
#>  $ CCC_072 : int  2 6 2 2 2 2 6 6 2 2 ...
#>  $ CCC_073 : int  6 1 6 6 6 6 1 1 6 6 ...
#>  $ CCC_073A: int  6 2 6 6 6 6 6 2 6 6 ...
#>  $ CCC_073B: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_081 : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ CCC_091 : int  2 2 2 6 2 2 2 2 2 6 ...
#>  $ CCC_101 : int  2 1 1 2 2 2 2 2 2 2 ...
#>  $ CCCG102 : int  96 9 11 96 96 96 96 96 96 96 ...
#>  $ CCC_10A : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_10B : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_10C : int  96 6 6 96 96 96 96 96 96 96 ...
#>  $ CCC_105 : int  6 2 2 6 6 6 6 6 6 6 ...
#>  $ CCC_106 : int  6 1 1 6 6 6 6 6 6 6 ...
#>   [list output truncated]