Long term risks due to drinking

This function creates a categorical variable that flags for increased long term health risks due to their drinking habits, according to Canada's Low-Risk Alcohol Drinking Guideline.

low_drink_long_fun(...)

Arguments

DHH_SEX

Sex of respondent (1 - male, 2 - female)

ALW_1

Drinks in the last week (1 - yes, 2 - no)

ALC_1

Drinks in the past year (1 - yes, 2 - no)

ALW_2A1

Number of drinks on Sunday

ALW_2A2

Number of drinks on Monday

ALW_2A3

Number of drinks on Tuesday

ALW_2A4

Number of drinks on Wednesday

ALW_2A5

Number of drinks on Thursday

ALW_2A6

Number of drinks on Friday

ALW_2A7

Number of drinks on Saturday

ALWDWKY

Number of drinks consumed in the past week

Value

Categorical variable (ALWDVLTR_der) with two categories:

1 - Increased long term health risk

2 - No increased long term health risk

Details

The classification of drinkers according to their long term health risks comes from guidelines in Alcohol and Health in Canada: A Summary of Evidence and Guidelines for Low-risk Drinking, and is based on the alcohol consumption reported over the past week. Short-term or acute risks include injury and overdose.

Categories are based on CCHS 2015-2016's variable (ALWDVLTR) where long term health risk are increased when drinking more than 10 drinks a week for women, with no more than 2 drinks a day most days, and more than 15 drinks a week for men, with no more than 3 drinks a day most days.

See https://osf.io/ykau5/ for more details on the guideline. See https://osf.io/ycxaq/ for more details on the derivation of the function on page 8.

Examples