Smoking pack-years

This function creates a derived variable (pack_years_der) that measures an individual's smoking pack-years based on various CCHS smoking variables. This is a popular variable used by researchers to quantify lifetime exposure to cigarette use.

pack_years_fun(
  SMKDSTY,
  DHHGAGE_cont,
  SMK_09A_B,
  SMKG09C,
  SMKG203_cont,
  SMKG207_cont,
  SMK_204,
  SMK_05B,
  SMK_208,
  SMK_05C,
  SMKG01C_cont,
  SMK_01A
)

Arguments

SMKDSTY
derived variable that classifies an individual's smoking status.

DHHGAGE_cont
continuous age variable.

SMK_09A_B
number of years since quitting smoking. Variable asked to former daily smokers who quit <3 years ago.

SMKG09C
number of years since quitting smoking. Variable asked to former daily smokers who quit >=3 years ago.

SMKG203_cont
age started smoking daily. Variable asked to daily smokers.

SMKG207_cont
age started smoking daily. Variable asked to former daily smokers.

SMK_204
number of cigarettes smoked per day. Variable asked to daily smokers.

SMK_05B
number of cigarettes smoked per day. Variable asked to occasional smokers

SMK_208
number of cigarettes smoked per day. Variable asked to former daily smokers

SMK_05C
number of days smoked at least one cigarette

SMKG01C_cont
age smoked first cigarette

SMK_01A
smoked 100 cigarettes in lifetime (y/n)

Value

value for smoking pack-years in the Pack_years_der variable

Details

pack-years is calculated by multiplying the number of cigarette packs per day (20 cigarettes per pack) by the number of years. Example 1: a respondent who is a current smoker who smokes 1 package of cigarettes for the last 10 years has smoked 10 pack-years. Pack-years is also calculated for former smokers. Example 2: a respondent who started smoking at age 20 years and smoked half a pack of cigarettes until age 40 years smoked for 10 pack-years.

Examples

# Using pack_years_fun() to create pack-years values across CCHS cycles
# pack_years_fun() is specified in variable_details.csv along with the CCHS
# variables and cycles included.

# To transform pack_years_der across cycles, use rec_with_table() for each
# CCHS cycle and specify pack_years_der, along with each smoking variable.
# Then by using bind_rows(), you can combine pack_years_der across cycles

library(cchsflow)

pack_years2010 <- rec_with_table(
  cchs2010, c(
    "SMKDSTY", "DHHGAGE_cont", "SMK_09A_B", "SMKG09C",
    "SMKG203_cont", "SMKG207_cont", "SMK_204", "SMK_05B", "SMK_208",
    "SMK_05C", "SMK_01A", "SMKG01C_cont", "pack_years_der"
  )
)

#> Loading cchsflow variable_details

#> Using the passed data variable name as database_name

head(pack_years2010)

#>   DHHGAGE_cont SMK_01A SMK_05B SMK_05C SMK_09A_B SMK_204 SMK_208 SMKDSTY
#> 1           22       2      NA      NA     NA(a)      NA      NA       6
#> 2           42       2      NA      NA     NA(a)      NA      NA       6
#> 3           37       1      NA      NA     NA(a)      10      NA       1
#> 4           37       2      NA      NA     NA(a)      NA      NA       6
#> 5           37       2      NA      NA     NA(a)      NA      NA       6
#> 6           32       1      NA      NA         1      NA      12       4
#>   SMKG01C_cont SMKG09C SMKG203_cont SMKG207_cont pack_years_der
#> 1           NA   NA(a)           NA           NA           0.00
#> 2           NA   NA(a)           NA           NA           0.00
#> 3         18.5   NA(a)         18.5           NA           9.25
#> 4           NA   NA(a)           NA           NA           0.00
#> 5           NA   NA(a)           NA           NA           0.00
#> 6         16.0   NA(a)           NA           16           9.30

pack_years2012 <- rec_with_table(
  cchs2012,c(
    "SMKDSTY", "DHHGAGE_cont", "SMK_09A_B", "SMKG09C",
    "SMKG203_cont", "SMKG207_cont", "SMK_204", "SMK_05B", "SMK_208",
    "SMK_05C", "SMK_01A", "SMKG01C_cont", "pack_years_der"
  )
)

#> Loading cchsflow variable_details

#> Using the passed data variable name as database_name

tail(pack_years2012)

#>     DHHGAGE_cont SMK_01A SMK_05B SMK_05C SMK_09A_B SMK_204 SMK_208 SMKDSTY
#> 195           42       2      NA      NA     NA(a)      NA      NA       6
#> 196           57       2      NA      NA     NA(a)      NA      NA       6
#> 197           32       1      NA      NA         3      NA      20       4
#> 198           77       2      NA      NA     NA(a)      NA      NA       5
#> 199           47       1      NA      NA     NA(a)       5      NA       1
#> 200           85       2      NA      NA     NA(a)      NA      NA       5
#>     SMKG01C_cont SMKG09C SMKG203_cont SMKG207_cont pack_years_der
#> 195           NA   NA(a)           NA           NA          0.000
#> 196           NA   NA(a)           NA           NA          0.000
#> 197          8.0   NA(a)           NA           13         16.500
#> 198         18.5   NA(a)           NA           NA          0.007
#> 199         16.0   NA(a)           16           NA          7.750
#> 200          8.0   NA(a)           NA           NA          0.007

combined_pack_years <- bind_rows(pack_years2010, pack_years2012)

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: Unequal factor levels: coercing to character

#> Warning: binding character and factor vector, coercing into character vector

#> Warning: binding character and factor vector, coercing into character vector

head(combined_pack_years)

#>   DHHGAGE_cont SMK_01A SMK_05B SMK_05C SMK_09A_B SMK_204 SMK_208 SMKDSTY
#> 1           22       2      NA      NA     NA(a)      NA      NA       6
#> 2           42       2      NA      NA     NA(a)      NA      NA       6
#> 3           37       1      NA      NA     NA(a)      10      NA       1
#> 4           37       2      NA      NA     NA(a)      NA      NA       6
#> 5           37       2      NA      NA     NA(a)      NA      NA       6
#> 6           32       1      NA      NA         1      NA      12       4
#>   SMKG01C_cont SMKG09C SMKG203_cont SMKG207_cont pack_years_der
#> 1           NA   NA(a)           NA           NA           0.00
#> 2           NA   NA(a)           NA           NA           0.00
#> 3         18.5   NA(a)         18.5           NA           9.25
#> 4           NA   NA(a)           NA           NA           0.00
#> 5           NA   NA(a)           NA           NA           0.00
#> 6         16.0   NA(a)           NA           16           9.30

tail(combined_pack_years)

#>     DHHGAGE_cont SMK_01A SMK_05B SMK_05C SMK_09A_B SMK_204 SMK_208 SMKDSTY
#> 395           42       2      NA      NA     NA(a)      NA      NA       6
#> 396           57       2      NA      NA     NA(a)      NA      NA       6
#> 397           32       1      NA      NA         3      NA      20       4
#> 398           77       2      NA      NA     NA(a)      NA      NA       5
#> 399           47       1      NA      NA     NA(a)       5      NA       1
#> 400           85       2      NA      NA     NA(a)      NA      NA       5
#>     SMKG01C_cont SMKG09C SMKG203_cont SMKG207_cont pack_years_der
#> 395           NA   NA(a)           NA           NA          0.000
#> 396           NA   NA(a)           NA           NA          0.000
#> 397          8.0   NA(a)           NA           13         16.500
#> 398         18.5   NA(a)           NA           NA          0.007
#> 399         16.0   NA(a)           16           NA          7.750
#> 400          8.0   NA(a)           NA           NA          0.007