Percent time in Canada

This function creates a derived variable (pct_time_der) that provides an estimated percentage of the time a person's life was spent in Canada.

pct_time_fun(DHHGAGE_cont, SDCGCBG, SDCGRES)

Arguments

DHHGAGE_cont
continuous age variable.

SDCGCBG
whether or not someone was born in Canada (1 - born in Canada, 2 - born outside Canada)

SDCGRES
how long someone has lived in Canada. Note: in the PUMF CCHS datasets, this is a categorical variable with two categories (1 - 0-9 years; 2 - 10+ years).

Value

Numeric value that is a fraction between 0 and 1 that represents percentage of a respondent's time in Canada

Note

Since SDCGRES is a categorical variable measuring length of time, we've set midpoints in the function. A respondent identified as being in Canada for 0-9 years is assigned a value of 4.5 years, and someone who has been in Canada for over 10 years is assigned a value of 15 years.

Examples

# Using pct_time_fun() to create percent time values between CCHS cycles
# pct_time_fun() is specified in variable_details.csv along with the CCHS
# variables and cycles included.

# To transform pct_time_der across cycles, use rec_with_table() for each CCHS
# cycle and specify pct_time_der, along with age (DHHGAGE_cont), whether or
# not someone was born in Canada (SDCGCBG), how long someone has lived in
# Canada (SDCGRES). Then by using bind_rows(), you can combine Pct_time_der
# across cycles

library(cchsflow)
pct_time2010 <- rec_with_table(
  cchs2010_p, c(
    "DHHGAGE_cont", "SDCGCBG",
    "SDCGRES", "pct_time_der"
  )
)

#> Loading cchsflow variable_details

#> Using the passed data variable name as database_name

#> NOTE for SDCGCBG: CCHS 2001 does not have don't know (7) or refusal (8)

#> NOTE for SDCGRES: CCHS 2001 missing don't know (7), refusal (8)

head(pct_time2010)

#>   DHHGAGE_cont SDCGCBG SDCGRES pct_time_der
#> 1           22       1   NA(a)    1.0000000
#> 2           42       2       2    0.3571429
#> 3           37       1   NA(a)    1.0000000
#> 4           37       1   NA(a)    1.0000000
#> 5           37       1   NA(a)    1.0000000
#> 6           32       1   NA(a)    1.0000000

pct_time2012 <- rec_with_table(
  cchs2012_p,  c(
    "DHHGAGE_cont", "SDCGCBG",
    "SDCGRES", "pct_time_der"
  )
)

#> Loading cchsflow variable_details

#> Using the passed data variable name as database_name

#> NOTE for SDCGCBG: CCHS 2001 does not have don't know (7) or refusal (8)

#> NOTE for SDCGRES: CCHS 2001 missing don't know (7), refusal (8)

tail(pct_time2012)

#>     DHHGAGE_cont SDCGCBG SDCGRES pct_time_der
#> 195           42       1   NA(a)     1.000000
#> 196           57       1   NA(a)     1.000000
#> 197           32       2       1     0.140625
#> 198           77       1   NA(a)     1.000000
#> 199           47       1   NA(a)     1.000000
#> 200           85       1   NA(a)     1.000000

combined_pct_time <- bind_rows(pct_time2010, pct_time2012)
head(combined_pct_time)

#>   DHHGAGE_cont SDCGCBG SDCGRES pct_time_der
#> 1           22       1   NA(a)    1.0000000
#> 2           42       2       2    0.3571429
#> 3           37       1   NA(a)    1.0000000
#> 4           37       1   NA(a)    1.0000000
#> 5           37       1   NA(a)    1.0000000
#> 6           32       1   NA(a)    1.0000000

tail(combined_pct_time)

#>     DHHGAGE_cont SDCGCBG SDCGRES pct_time_der
#> 395           42       1   NA(a)     1.000000
#> 396           57       1   NA(a)     1.000000
#> 397           32       2       1     0.140625
#> 398           77       1   NA(a)     1.000000
#> 399           47       1   NA(a)     1.000000
#> 400           85       1   NA(a)     1.000000

# Using pct_time_fun() to generate a value for percent time spent in Canada
# with user inputted values Let's say you are 27 years old who was born
# outside of Canada and have been living in Canada for less than 10 years.
# Your estimated percent time spent in Canada can be calculated as follows:

pct_time <- pct_time_fun(DHHGAGE_cont = 27, SDCGCBG = 2, SDCGRES = 1)

print(pct_time)

#> [1] 0.1666667