resp_condition_fun3

This is one of 3 functions used to create a derived variable (resp_condition_der) that determines if a respondents has a respiratory condition. This function for CCHS cycles (2001-2003) that use COPD and Emphysema as a combined variable, as well as Bronchitis. Asthma is used across CCHS cycles as a separate variable.

resp_condition_fun3(DHHGAGE_cont, CCC_091, CCC_91A, CCC_031)

Arguments

DHHGAGE_cont
continuous age variable.

CCC_091
variable indicating if respondent has either COPD or Emphysema

CCC_91A
variable indicating if respondent has chronic bronchitis

CCC_031
variable indicating if respondent has asthma

Value

a categorical variable (resp_condition_der) with 3 levels:

respondent is over the age of 35 and has a respiratory condition

respondent is under the age of 35 and has a respiratory condition

respondent does not have a respiratory condition

See also

Examples

# Using resp_condition_fun3() to create values across CCHS cycles
# (2001-2003) resp_condition_fun3() is specified in
# variable_details.csv along with the CCHS variables and cycles included.

# To transform resp_condition_der, use rec_with_table() for each CCHS cycle
# and specify resp_condition_der, along with the various respiratory
# variables. Then by using bind_rows() you can combine resp_condition_der
# across cycles.

library(cchsflow)

resp2001 <- suppressWarnings(rec_with_table(
  cchs2001_p, c(
    "DHHGAGE_cont", "CCC_091", "CCC_91A", "CCC_031",
    "resp_condition_der"
  )
))

#> No variable_details detected.
#>               Loading cchsflow variable_details

#> Using the passed data variable name as database_name

#> NOTE for DHHGAGE_cont: Not applicable, don't know, refusal, not stated (96-99) were options in CCHS 2003, but had zero responses

head(resp2001)

#>   CCC_031 CCC_091 CCC_91A DHHGAGE_cont resp_condition_der
#> 1       2       2       2           85                  3
#> 2       2       2       2           85                  3
#> 3       2       2       2           32                  3
#> 4       2       2       2           77                  3
#> 5       2   NA(a)       2           22                  3
#> 6       2       2       2           77                  3

resp2003 <- suppressWarnings(rec_with_table(
  cchs2003_p,c(
    "DHHGAGE_cont", "CCC_091", "CCC_91A", "CCC_031",
    "resp_condition_der"
  )
))

#> No variable_details detected.
#>               Loading cchsflow variable_details

#> Using the passed data variable name as database_name

#> NOTE for DHHGAGE_cont: Not applicable, don't know, refusal, not stated (96-99) were options in CCHS 2003, but had zero responses

tail(resp2003)

#>     CCC_031 CCC_091 CCC_91A DHHGAGE_cont resp_condition_der
#> 195       2   NA(a)       2           22                  3
#> 196       1       2       1           57                  1
#> 197       2       2       2           42                  3
#> 198       2       2       2           85                  3
#> 199       1   NA(a)       2           17                  2
#> 200       2   NA(a)       2           17                  3

combined_resp <- suppressWarnings(bind_rows(resp2001, resp2003))

head(combined_resp)

#>   CCC_031 CCC_091 CCC_91A DHHGAGE_cont resp_condition_der
#> 1       2       2       2           85                  3
#> 2       2       2       2           85                  3
#> 3       2       2       2           32                  3
#> 4       2       2       2           77                  3
#> 5       2   NA(a)       2           22                  3
#> 6       2       2       2           77                  3

tail(combined_resp)

#>     CCC_031 CCC_091 CCC_91A DHHGAGE_cont resp_condition_der
#> 395       2   NA(a)       2           22                  3
#> 396       1       2       1           57                  1
#> 397       2       2       2           42                  3
#> 398       2       2       2           85                  3
#> 399       1   NA(a)       2           17                  2
#> 400       2   NA(a)       2           17                  3