variable_details.csv

This dataset provides details on how variables are recoded in cchsflow.

Value

variable_details

a data frame

Details

See the below link for more details about how the worksheet is structured https://big-life-lab.github.io/cchsflow/articles/variable_details.html

Examples

data(variable_details)
str(variable_details)

#> 'data.frame':	1172 obs. of  16 variables:
#>  $ variable               : Factor w/ 150 levels "ADL_01","ADL_02",..: 1 1 1 1 1 2 2 2 2 2 ...
#>  $ dummyVariable          : Factor w/ 672 levels "ADL_01_cat2_1",..: 1 2 3 4 4 5 6 7 8 8 ...
#>  $ toType                 : Factor w/ 3 levels "cat","cont","N/A": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ databaseStart          : Factor w/ 31 levels "cchs-82M0013-E-2005-c3-1-main-file, cchs-E-2007-2008-AnnualComponent, CCHS-82M0013-E-2009-2010-Annualcomponent,"| __truncated__,..: 6 6 6 6 6 17 17 17 17 17 ...
#>  $ variableStart          : Factor w/ 151 levels "[ALCDTTM]","[FSCDHFS2]",..: 86 86 86 86 86 120 120 120 120 120 ...
#>  $ fromType               : Factor w/ 3 levels "cat","cont","N/A": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ recTo                  : Factor w/ 98 levels "0","0.5","1",..: 3 49 97 98 98 3 49 97 98 98 ...
#>  $ numValidCat            : Factor w/ 11 levels "10","11","12",..: 6 6 6 6 6 6 6 6 6 6 ...
#>  $ catLabel               : Factor w/ 292 levels "# days smoked at least 1 cigarette",..: 291 202 238 185 145 291 202 238 185 145 ...
#>  $ catLabelLong           : Factor w/ 397 levels "# days smoked at least 1 cigarette",..: 395 297 336 276 236 395 297 336 276 236 ...
#>  $ units                  : Factor w/ 19 levels "%","$/year","cigarettes",..: 16 16 16 16 16 16 16 16 16 16 ...
#>  $ recFrom                : Factor w/ 78 levels "-0.359:1","0",..: 11 29 55 59 77 11 29 55 59 77 ...
#>  $ catStartLabel          : Factor w/ 344 levels "# days smoked at least 1 cigarette",..: 324 224 260 161 173 324 224 260 161 173 ...
#>  $ variableStartShortLabel: Factor w/ 133 levels "Act. prevent/pain - function code - (D)",..: 93 93 93 93 93 89 89 89 89 89 ...
#>  $ variableStartLabel     : Factor w/ 131 levels "[Other than the ( X ) times you already reported bicycling was there any other time / Was there any other time]"| __truncated__,..: 17 17 17 17 17 18 18 18 18 18 ...
#>  $ notes                  : Factor w/ 40 levels "","2001 and 2003 CCHS use inches, values converted to meters to 3 decimal points",..: 1 1 1 1 1 1 1 1 1 1 ...