variables.csv

This dataset lists all the variables that are present in cchsflow.

Value

variables

a data frame

Details

See the below link for more details about how the worksheet is structured https://big-life-lab.github.io/cchsflow/articles/variables_sheet.html

Examples

data(variables)
str(variables)

#> 'data.frame':	150 obs. of  9 variables:
#>  $ variable     : Factor w/ 150 levels "ADL_01","ADL_02",..: 1 2 3 4 5 6 7 8 9 10 ...
#>  $ label        : Factor w/ 128 levels "# of drinks - Friday",..: 59 54 55 57 56 58 60 15 41 40 ...
#>  $ labelLong    : Factor w/ 130 levels "# days smoked at least 1 cigarette",..: 82 78 77 81 80 79 83 96 119 118 ...
#>  $ section      : Factor w/ 6 levels "Demographics",..: 4 4 4 4 4 4 4 2 2 2 ...
#>  $ subject      : Factor w/ 29 levels "ADL","Age","Alcohol",..: 1 1 1 1 1 1 1 3 3 3 ...
#>  $ variableType : Factor w/ 2 levels "Categorical",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ units        : Factor w/ 16 levels "%","$/year","cigarettes",..: 14 14 14 14 14 14 14 14 14 14 ...
#>  $ databaseStart: Factor w/ 23 levels "cchs2001_p","cchs2001_p, cchs2003_p",..: 5 16 5 5 5 16 5 5 5 4 ...
#>  $ variableStart: Factor w/ 150 levels "[FSCDHFS2]","[FSCEDHFS]",..: 87 124 88 89 90 125 92 18 28 29 ...