2012 CCHS synthetic subset data (200 respondents)

This is a subset of 200 observations from the 2012 cycle of the Canadian Community Health Survey (CCHS) synthetic dataset. The CCHS survey is conducted by Statistics Canada.

Source

https://www.statcan.gc.ca/en/statistical-programs/document/3226_D56_T9_V1

Value

cchs2012_s

a data frame

Details

NOTE: this subset of respondents may also be in the 2012 synthetic subset. Please see the "CCHS datasets that overlap each other" article to see how the two datasets contain overlap.

See here for the open license. Source from Statistics Canada, Canadian Community Health Survey PUMF, accessed May 2022. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.

Long name: synthetic-CCHS-E-2012-AnnualComponent_F1

DDI: https://osf.io/q8g7y/

Additional documentation (PDFs): https://osf.io/hkuy3/

Examples

data(cchs2012_s)
str(cchs2012_s)
#> 'data.frame':	200 obs. of  1815 variables:
#>  $ VERDATE : int  20130606 20130606 20130606 20130606 20130606 20130606 20130606 20130606 20130606 20130606 ...
#>  $ REFPER  : Factor w/ 1 level "201201-201212": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ SAMPLEID: int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ PERSONID: logi  NA NA NA NA NA NA ...
#>  $ GEO_PRV : int  10 10 10 10 10 10 10 10 10 10 ...
#>  $ GEODPC  : Factor w/ 200 levels "A0G1E0","A0G1R0",..: 171 122 1 143 74 114 161 158 6 153 ...
#>  $ GEODHR4 : int  1014 1012 1012 1013 1011 1012 1014 1013 1012 1013 ...
#>  $ GEODSHR : int  999996 999996 999996 999996 999996 999996 999996 999996 999996 999996 ...
#>  $ GEODDHA : int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODLHA : int  996 996 996 996 996 996 996 996 996 996 ...
#>  $ GEODRHA : int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODBCHA: int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODLHN : int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ GEODDA06: int  10100090 10060112 10080244 10050144 10010647 10060105 10100101 10040083 10080184 10040088 ...
#>  $ GEODFED : int  10004 10002 10002 10003 10007 10002 10004 10005 10002 10005 ...
#>  $ GEODCSD : int  1010032 1006017 1008038 1005018 1001542 1006017 1010032 1004019 1008033 1004019 ...
#>  $ GEODCD  : int  1010 1006 1008 1005 1001 1006 1010 1004 1008 1004 ...
#>  $ GEODSAT : int  6 3 7 3 1 3 6 6 7 6 ...
#>  $ GEODCMA6: int  0 0 0 0 1 0 0 0 0 0 ...
#>  $ GEODPG11: int  8 9 9 9 3 9 8 9 9 9 ...
#>  $ GEODUR  : int  4 9 9 1 9 1 4 4 0 4 ...
#>  $ GEODUR2 : int  1 1 1 1 1 1 1 1 2 1 ...
#>  $ GEODSUBZ: int  999996 999996 999996 999996 999996 999996 999996 999996 999996 999996 ...
#>  $ GEODRMB : int  9996 9996 9996 9996 9996 9996 9996 9996 9996 9996 ...
#>  $ SAM_CP  : logi  NA NA NA NA NA NA ...
#>  $ SAM_TYP : logi  NA NA NA NA NA NA ...
#>  $ SAMDSHR : logi  NA NA NA NA NA NA ...
#>  $ SAMDLNK : logi  NA NA NA NA NA NA ...
#>  $ ADM_STA : logi  NA NA NA NA NA NA ...
#>  $ ADM_PRX : logi  NA NA NA NA NA NA ...
#>  $ ADM_YOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_MOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_DOI : logi  NA NA NA NA NA NA ...
#>  $ ADM_N09 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N10 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N11 : logi  NA NA NA NA NA NA ...
#>  $ ADM_N12 : logi  NA NA NA NA NA NA ...
#>  $ ADM_LHH : logi  NA NA NA NA NA NA ...
#>  $ DHH_AGE : int  61 72 68 86 31 79 56 41 72 81 ...
#>  $ DHH_YOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_MOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_DOB : logi  NA NA NA NA NA NA ...
#>  $ DHH_SEX : int  1 1 2 2 2 2 2 1 2 1 ...
#>  $ DHH_MS  : int  1 1 1 3 6 1 1 1 3 6 ...
#>  $ DHHDSAGE: int  63 67 996 78 28 996 60 40 996 996 ...
#>  $ DHHDYKD : int  0 0 0 0 4 0 0 1 0 0 ...
#>  $ DHHDOKD : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHDLE5 : int  0 0 0 0 1 0 0 0 0 0 ...
#>  $ DHHD611 : int  0 0 0 0 1 0 0 1 0 0 ...
#>  $ DHHDL12 : int  0 0 0 0 2 0 0 1 0 0 ...
#>  $ DHHDL18 : int  0 0 0 0 2 0 0 1 0 0 ...
#>  $ DHHDLVG : int  3 3 1 3 10 1 3 4 1 1 ...
#>  $ DHHDECF : int  3 3 1 3 6 1 3 5 1 1 ...
#>  $ DHHDHSZ : int  2 2 1 2 10 1 2 3 1 1 ...
#>  $ GEN_01  : int  3 1 3 2 2 2 1 2 2 2 ...
#>  $ GEN_02  : int  3 3 3 3 2 2 3 3 3 3 ...
#>  $ GEN_02A2: int  9 99 7 10 9 8 9 10 10 99 ...
#>  $ GEN_02B : int  2 9 3 1 3 1 3 1 1 9 ...
#>  $ GEN_07  : int  4 2 1 1 4 2 4 3 2 2 ...
#>  $ GEN_08  : int  1 9 1 6 2 6 1 1 2 6 ...
#>  $ GEN_09  : int  4 9 2 6 6 6 1 3 6 6 ...
#>  $ GEN_10  : int  1 9 2 2 2 2 2 1 4 9 ...
#>  $ GENDHDI : int  2 4 2 3 3 3 4 3 3 3 ...
#>  $ GENDMHI : int  3 9 2 4 2 4 2 4 4 9 ...
#>  $ GENGSWL : int  1 9 2 1 1 2 1 1 1 9 ...
#>  $ DOORG   : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ ORG_1   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ ORG_2   : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ DOSLP   : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ SLP_01  : int  96 96 96 96 96 96 96 96 96 96 ...
#>  $ SLP_02  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ SLP_03  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ SLP_04  : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ DOCIH   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CIH_1   : int  2 2 1 2 1 1 2 2 1 2 ...
#>  $ CIH_2   : int  96 96 1 96 1 1 96 96 9 96 ...
#>  $ CIH_3   : int  2 1 1 2 2 1 1 1 1 1 ...
#>  $ CIH_4   : int  96 4 9 96 96 3 2 4 1 4 ...
#>  $ CIH_5   : int  6 1 1 6 6 2 1 2 1 2 ...
#>  $ CIH_6A  : int  6 1 2 6 6 6 1 6 1 6 ...
#>  $ CIH_6I  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6B  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6J  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6K  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6G  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6F  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6E  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6L  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6M  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6N  : int  6 2 2 6 6 6 2 6 2 6 ...
#>  $ CIH_6H  : int  6 2 1 6 6 6 2 6 2 6 ...
#>  $ CIH_7   : int  6 1 1 6 6 2 2 1 1 1 ...
#>  $ CIH_8A  : int  6 2 1 6 6 6 6 1 2 2 ...
#>  $ CIH_8B  : int  6 2 2 6 6 6 6 2 2 2 ...
#>  $ CIH_8C  : int  6 2 2 6 6 6 6 2 1 2 ...
#>  $ CIH_8J  : int  6 1 2 6 6 6 6 2 2 1 ...
#>  $ CIH_8K  : int  6 2 2 6 6 6 6 2 2 2 ...
#>  $ CIH_8G  : int  6 2 2 6 6 6 6 2 2 2 ...
#>  $ CIH_8L  : int  6 2 2 6 6 6 6 2 2 2 ...
#>   [list output truncated]