2017-2018 CCHS PUMF subset data (200 respondents)

This is a subset of 200 observations from the 2017-2018 cycle of the Canadian Community Health Survey (CCHS) Public Use Microdata file (PUMF) dataset. The CCHS survey is conducted by Statistics Canada.

Source

https://www23.statcan.gc.ca/imdb/p2SV.pl?Function=getSurvey&Id=329241

Value

cchs2017_2018_p

a data frame

Details

NOTE: this subset of respondents may also be in the 2017-2018 PUMF subset. Please see the "CCHS datasets that overlap each other" article to see how the two datasets contain overlap.

See here for the open license. Source from Statistics Canada, Canadian Community Health Survey PUMF, accessed Oct 2021. Reproduced and distributed on an "as is" basis with the permission of Statistics Canada.

Long name: cchs-82M0013-E-2017-2018-Annual-component

DDI: https://osf.io/q8g7y/

Additional documentation (PDFs): https://osf.io/hkuy3/

Examples

data(cchs2017_2018_p)
str(cchs2017_2018_p)
#> 'data.frame':	200 obs. of  1051 variables:
#>  $ ADM_RNO : int  21 29 39 54 63 71 72 125 127 133 ...
#>  $ VERDATE : int  20190913 20190913 20190913 20190913 20190913 20190913 20190913 20190913 20190913 20190913 ...
#>  $ GEO_PRV : int  46 48 13 59 35 24 46 24 35 35 ...
#>  $ GEODGHR4: int  46903 48931 13904 59931 35944 24916 46902 24903 35947 35953 ...
#>  $ GEODVBHA: int  9996 9996 9996 5930 9996 9996 9996 9996 9996 9996 ...
#>  $ DHH_SEX : int  2 1 2 2 2 2 1 2 2 2 ...
#>  $ DHHGMS  : int  3 3 1 3 1 2 2 2 2 1 ...
#>  $ DHHDGHSZ: int  1 1 4 1 5 3 4 4 4 3 ...
#>  $ DHHDGLVG: int  1 1 8 1 4 4 8 4 4 4 ...
#>  $ DHHDG611: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ DHHDGL12: int  0 0 0 0 0 1 0 0 1 0 ...
#>  $ DHHDGLE5: int  0 0 0 0 0 1 0 0 1 0 ...
#>  $ ADM_PRX : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ DHHGAGE : int  16 14 9 10 9 6 14 9 6 10 ...
#>  $ DOMAC   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ MACG005 : int  4 4 1 1 1 6 1 1 6 1 ...
#>  $ MAC_010 : int  6 2 1 1 1 2 1 1 1 1 ...
#>  $ MAC_015 : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ MACG020 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ EHG2DVR3: int  3 1 3 3 3 3 3 3 3 3 ...
#>  $ EHG2DVH3: int  3 1 3 3 3 3 3 3 3 3 ...
#>  $ DOGEN   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ GEN_005 : int  2 5 2 2 1 1 4 2 2 2 ...
#>  $ GEN_010 : int  10 10 8 10 10 9 8 8 8 6 ...
#>  $ GEN_015 : int  2 3 3 2 1 1 3 2 2 3 ...
#>  $ GEN_020 : int  3 1 4 1 3 1 3 2 3 3 ...
#>  $ GEN_025 : int  6 6 5 2 3 6 3 3 4 3 ...
#>  $ GEN_030 : int  2 4 1 3 2 2 2 3 2 3 ...
#>  $ GENDVHDI: int  3 0 3 3 4 4 1 3 3 3 ...
#>  $ GENDVMHI: int  3 2 2 3 4 4 2 3 3 2 ...
#>  $ GENDVSWL: int  1 1 2 1 1 1 2 2 2 2 ...
#>  $ DOHWT   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ HWT_050 : int  1 1 3 1 3 3 3 3 3 1 ...
#>  $ HWTDGHTM: num  1.65 1.73 1.6 1.68 1.75 ...
#>  $ HWTDGWTK: num  74.2 108 60.8 81 63 ...
#>  $ HWTDGBMI: num  27.2 36.2 23.7 28.8 20.5 ...
#>  $ HWTDVCOL: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ HWTDGISW: int  3 4 2 3 2 2 3 2 2 2 ...
#>  $ HWTDVWHO: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ HWTDGCOR: num  28.5 37.9 24.8 30.2 21.4 ...
#>  $ HWTDGBCC: int  3 4 2 4 2 2 3 2 2 2 ...
#>  $ DOCCC   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ CCC_005 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_010 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_015 : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ CCC_020 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_025 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_030 : int  2 1 2 2 2 6 2 2 6 2 ...
#>  $ CCC_035 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_040 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_045 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_050 : int  2 2 1 1 2 2 2 2 2 2 ...
#>  $ CCC_055 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_060 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_065 : int  1 1 2 1 2 2 1 2 2 2 ...
#>  $ CCC_070 : int  1 1 2 1 2 2 1 2 2 2 ...
#>  $ CCC_075 : int  2 2 2 1 2 2 1 2 2 2 ...
#>  $ CCC_080 : int  2 7 2 1 2 2 1 2 2 2 ...
#>  $ CCC_085 : int  2 1 2 2 2 2 1 2 2 2 ...
#>  $ CCC_090 : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ CCC_095 : int  2 1 2 2 2 2 2 2 2 2 ...
#>  $ CCCG100 : int  96 13 96 96 96 96 96 96 96 96 ...
#>  $ CCC_105 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_110 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_115 : int  96 6 96 96 96 96 96 96 96 96 ...
#>  $ CCC_120 : int  6 2 6 6 6 6 6 6 6 6 ...
#>  $ CCC_125 : int  6 1 6 6 6 6 6 6 6 6 ...
#>  $ CCC_130 : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ CCC_135 : int  2 1 2 2 2 2 2 2 2 2 ...
#>  $ CCC_140 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_170 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_175A: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_175B: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_175C: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_175D: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_175E: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_175F: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_175G: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_175H: int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_180 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_185 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_190 : int  6 6 6 6 6 6 6 6 6 6 ...
#>  $ CCC_195 : int  1 2 2 2 2 2 2 2 2 2 ...
#>  $ CCC_200 : int  2 2 1 2 2 2 2 2 2 2 ...
#>  $ CCCDVDIA: int  6 2 6 6 6 6 6 6 6 6 ...
#>  $ DOWDM   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ WDM_005 : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ WDM_010 : int  2 1 1 1 1 1 2 1 1 1 ...
#>  $ WDM_015 : int  2 2 1 1 1 1 1 1 1 1 ...
#>  $ WDM_020 : int  1 2 1 1 1 1 2 1 1 1 ...
#>  $ WDM_025 : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ WDM_030 : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ DOADL   : int  2 2 1 2 2 1 2 1 2 2 ...
#>  $ ADL_005 : int  6 6 1 6 6 1 6 1 6 6 ...
#>  $ ADL_010 : int  6 6 1 6 6 1 6 1 6 6 ...
#>  $ ADL_015 : int  6 6 1 6 6 1 6 1 6 6 ...
#>  $ ADL_020 : int  6 6 1 6 6 1 6 1 6 6 ...
#>  $ ADL_025 : int  6 6 1 6 6 1 6 1 6 6 ...
#>  $ ADL_030 : int  6 6 1 6 6 1 6 1 6 6 ...
#>   [list output truncated]