Smoking pack-years

This function creates a derived variable (pack_years_der) that measures an individual's smoking pack-years based on various CCHS smoking variables. This is a popular variable used by researchers to quantify lifetime exposure to cigarette use.

pack_years_fun(
  SMKDSTY,
  DHHGAGE_cont,
  SMK_09A_B,
  SMKG09C,
  SMKG203_cont,
  SMKG207_cont,
  SMK_204,
  SMK_05B,
  SMK_208,
  SMK_05C,
  SMKG01C_cont,
  SMK_01A
)

Arguments

SMKDSTY
derived variable that classifies an individual's smoking status.

DHHGAGE_cont
continuous age variable.

SMK_09A_B
number of years since quitting smoking. Variable asked to former daily smokers who quit <3 years ago.

SMKG09C
number of years since quitting smoking. Variable asked to former daily smokers who quit >=3 years ago.

SMKG203_cont
age started smoking daily. Variable asked to daily smokers.

SMKG207_cont
age started smoking daily. Variable asked to former daily smokers.

SMK_204
number of cigarettes smoked per day. Variable asked to daily smokers.

SMK_05B
number of cigarettes smoked per day. Variable asked to occasional smokers

SMK_208
number of cigarettes smoked per day. Variable asked to former daily smokers

SMK_05C
number of days smoked at least one cigarette

SMKG01C_cont
age smoked first cigarette

SMK_01A
smoked 100 cigarettes in lifetime (y/n)

Value

value for smoking pack-years in the Pack_years_der variable

Details

pack-years is calculated by multiplying the number of cigarette packs per day (20 cigarettes per pack) by the number of years. Example 1: a respondent who is a current smoker who smokes 1 package of cigarettes for the last 10 years has smoked 10 pack-years. Pack-years is also calculated for former smokers. Example 2: a respondent who started smoking at age 20 years and smoked half a pack of cigarettes until age 40 years smoked for 10 pack-years.

Examples

# Using pack_years_fun() to create pack-years values across CCHS cycles
# pack_years_fun() is specified in variable_details.csv along with the CCHS
# variables and cycles included.

# To transform pack_years_der across cycles, use rec_with_table() for each
# CCHS cycle and specify pack_years_der, along with each smoking variable.
# Then by using bind_rows(), you can combine pack_years_der across cycles

library(cchsflow)

pack_years2009_2010 <- rec_with_table(
  cchs2009_2010_p, c(
    "SMKDSTY", "DHHGAGE_cont", "SMK_09A_B", "SMKG09C",
    "SMKG203_cont", "SMKG207_cont", "SMK_204", "SMK_05B", "SMK_208",
    "SMK_05C", "SMK_01A", "SMKG01C_cont", "pack_years_der"
  )
)

#> Loading cchsflow variable_details

#> Using the passed data variable name as database_name

head(pack_years2009_2010)

#>   DHHGAGE_cont SMK_01A SMK_05B SMK_05C SMK_09A_B SMK_204 SMK_208 SMKDSTY
#> 1           13       2      NA      NA     NA(a)      NA      NA       6
#> 2           27       2      NA      NA     NA(a)      NA      NA       6
#> 3           62       2      NA      NA     NA(a)      NA      NA       5
#> 4           52       1      NA      NA         4      NA      20       4
#> 5           67       1      NA      NA         4      NA      35       4
#> 6           62       1      NA      NA         4      NA      25       4
#>   SMKG01C_cont SMKG09C SMKG203_cont SMKG207_cont pack_years_der
#> 1           NA   NA(a)           NA           NA          0.000
#> 2           NA   NA(a)           NA           NA          0.000
#> 3           16   NA(a)           NA           NA          0.007
#> 4           16       2           NA         18.5         25.500
#> 5           16       3           NA         16.0         68.250
#> 6            8       3           NA          8.0         52.500

pack_years2011_2012 <- rec_with_table(
  cchs2011_2012_p,c(
    "SMKDSTY", "DHHGAGE_cont", "SMK_09A_B", "SMKG09C",
    "SMKG203_cont", "SMKG207_cont", "SMK_204", "SMK_05B", "SMK_208",
    "SMK_05C", "SMK_01A", "SMKG01C_cont", "pack_years_der"
  )
)

#> Loading cchsflow variable_details

#> Using the passed data variable name as database_name

tail(pack_years2011_2012)

#>     DHHGAGE_cont SMK_01A SMK_05B SMK_05C SMK_09A_B SMK_204 SMK_208 SMKDSTY
#> 195         72.0       1      NA      NA         4      NA       3       4
#> 196         22.0       2      NA      NA     NA(a)      NA      NA       5
#> 197         18.5       2      NA      NA     NA(a)      NA      NA       6
#> 198         42.0       1      NA      NA         4      NA      10       4
#> 199         62.0       1      NA      NA     NA(a)      20      NA       1
#> 200         52.0       1      NA      NA     NA(a)      13      NA       1
#>     SMKG01C_cont SMKG09C SMKG203_cont SMKG207_cont pack_years_der
#> 195           16       3           NA           16          6.600
#> 196           13   NA(a)           NA           NA          0.007
#> 197           NA   NA(a)           NA           NA          0.000
#> 198           16       2           NA           16          9.000
#> 199           16   NA(a)           NA           NA             NA
#> 200           13   NA(a)           16           NA         23.400

combined_pack_years <- suppressWarnings(bind_rows(pack_years2009_2010,
 pack_years2011_2012))

head(combined_pack_years)

#>   DHHGAGE_cont SMK_01A SMK_05B SMK_05C SMK_09A_B SMK_204 SMK_208 SMKDSTY
#> 1           13       2      NA      NA     NA(a)      NA      NA       6
#> 2           27       2      NA      NA     NA(a)      NA      NA       6
#> 3           62       2      NA      NA     NA(a)      NA      NA       5
#> 4           52       1      NA      NA         4      NA      20       4
#> 5           67       1      NA      NA         4      NA      35       4
#> 6           62       1      NA      NA         4      NA      25       4
#>   SMKG01C_cont SMKG09C SMKG203_cont SMKG207_cont pack_years_der
#> 1           NA   NA(a)           NA           NA          0.000
#> 2           NA   NA(a)           NA           NA          0.000
#> 3           16   NA(a)           NA           NA          0.007
#> 4           16       2           NA         18.5         25.500
#> 5           16       3           NA         16.0         68.250
#> 6            8       3           NA          8.0         52.500

tail(combined_pack_years)

#>     DHHGAGE_cont SMK_01A SMK_05B SMK_05C SMK_09A_B SMK_204 SMK_208 SMKDSTY
#> 395         72.0       1      NA      NA         4      NA       3       4
#> 396         22.0       2      NA      NA     NA(a)      NA      NA       5
#> 397         18.5       2      NA      NA     NA(a)      NA      NA       6
#> 398         42.0       1      NA      NA         4      NA      10       4
#> 399         62.0       1      NA      NA     NA(a)      20      NA       1
#> 400         52.0       1      NA      NA     NA(a)      13      NA       1
#>     SMKG01C_cont SMKG09C SMKG203_cont SMKG207_cont pack_years_der
#> 395           16       3           NA           16          6.600
#> 396           13   NA(a)           NA           NA          0.007
#> 397           NA   NA(a)           NA           NA          0.000
#> 398           16       2           NA           16          9.000
#> 399           16   NA(a)           NA           NA             NA
#> 400           13   NA(a)           16           NA         23.400